Refactor Spark Connector into two packages
