+++ pinot-connectors/pinot-spark-2-connector/documentation/images/spark-pinot-connector-executor-server-interaction.jpg
