+++ pinot-connectors/pinot-spark-2-connector/documentation/read_model.md
