+++ pinot-connectors/pinot-spark-2-connector/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
