+++ pinot-connectors/pinot-spark-2-connector/src/test/resources/schema/pinot-schema.json
