+++ pinot-connectors/pinot-spark-common/src/test/resources/log4j2.xml
